fixing git probs ;(
