+++ test/cases/basic.json
@@ -37,7 +37,7 @@
           "title": {
             "en": "What's your name?",
             "fr": "some french"
-          },
+          }
         }
       ],
       "section": "personalDetails",
