--- test/cases/textfield.json
@@ -1,22 +1,0 @@
-{
-  "pages": [
-    {
-      "path": "/text-field",
-      "title": "TextField",
-      "components": [
-        {
-          "schema": {
-            "max": 70
-          },
-          "type": "TextField",
-          "name": "fullName",
-          "title": "Full name",
-          "hint": "Some help please...",
-          "options": {}
-        }
-      ]
-    }
-  ],
-  "sections": [],
-  "lists": []
-}
